requirements.txt
Flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.0.1

web3==6.20.1
eth-account==0.11.0

psycopg2-binary==2.9.9

opencv-python==4.9.0.80
numpy==1.26.4

qrcode==7.4.2
Pillow==10.2.0
